Re-enable sandbox cache and update AsyncLocalStorage usage
