--- packages/next/client/components/static-generation-async-storage/package.json
@@ -1,5 +1,0 @@
-{
-  "browser": {
-    "./storage.js": "./storage-browser.js"
-  }
-}
