--- packages/next/client/components/static-generation-async-storage/storage-browser.ts
@@ -1,13 +1,0 @@
-import type { AsyncLocalStorage } from 'async_hooks'
-
-export interface StaticGenerationStore {
-  inUse?: boolean
-  pathname?: string
-  revalidate?: number
-  fetchRevalidate?: number
-  isStaticGeneration?: boolean
-}
-
-export let staticGenerationAsyncStorage:
-  | AsyncLocalStorage<StaticGenerationStore>
-  | StaticGenerationStore = {}
