+++ packages/next/server/web/sandbox/context.ts
@@ -319,8 +319,15 @@ interface ModuleContextOptions {
   edgeFunctionEntry: Pick<EdgeFunctionDefinition, 'assets' | 'wasm'>
 }
 
+const pendingModuleCaches = new Map<string, Promise<ModuleContext>>()
+
 function getModuleContextShared(options: ModuleContextOptions) {
-  return createModuleContext(options)
+  let deferredModuleContext = pendingModuleCaches.get(options.moduleName)
+  if (!deferredModuleContext) {
+    deferredModuleContext = createModuleContext(options)
+    pendingModuleCaches.set(options.moduleName, deferredModuleContext)
+  }
+  return deferredModuleContext
 }
 
 /**
